refactor(agent): merge agent/definition sub-domain into definitions.ts
