--- packages/agent/src/agent/discovery/index.ts
@@ -1,7 +1,0 @@
-export { parseFrontmatter } from "./frontmatter";
-
-export {
-  AgentDiscovery,
-  type AgentLoadResult,
-  type AgentDiscoveryOptions,
-} from "./discovery";
