+++ packages/agent/src/agent/index.ts
@@ -1,6 +1,6 @@
 // Agent domain — identity, registry, discovery, messaging
 
-export * from "./definition";
+export * from "./definitions";
 export * from "./registry";
 export * from "./discovery";
 export * from "./communication";
