--- packages/agent/src/worker/policy/index.ts
@@ -1,15 +1,0 @@
-export { type ConcurrencyConfig, ConcurrencyGate } from "./concurrency";
-
-export {
-  type PermissionLevel,
-  type PermissionDecision,
-  type PermissionContext,
-  PermissionGate,
-} from "./permission";
-
-export {
-  type RunBudget,
-  type RunState,
-  type BudgetStatus,
-  RunSupervisor,
-} from "./run-supervisor";
